requirements.txt
# Core LangChain components
langchain>=0.0.267
langchain-community>=0.0.9
langchain-together>=0.0.3

# LLM Provider
together>=0.1.5

# Vector Database
chromadb>=0.4.18

# Embedding models
sentence-transformers>=2.2.2

# PDF Processing
pypdf>=3.15.1

# API and Web Server
flask>=2.0.1
flask-cors>=3.0.10

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.2

# Optional: For better search performance
faiss-cpu>=1.7.4

# HuggingFace dependencies
transformers>=4.30.0
torch>=2.0.0
